requirements.txt
django==4.0.4
django-cors-headers
django-cities-light
python-dateutil==2.8.2
Pillow
mysqlclient
reportlab
xlwt
django-admin-interface
django-notifications-hq
django-jsignature
requests==2.26.0
django-import-export
pandas
python-dotenv
django-crontab
matplotlib
django-funky-sheets